Super-quick go at generating button element submit
This work sparked an idea, buttons could be enabled in an ad hoc way by
allowing `button: true` to be set when calling `#govuk_submit`.

Like we've done elsewhere, `button` could take a proc and be called like
this:

= govuk_submit button: -> { tag.strong("Click me") }

This gives us loads of flexibility without adding too much complication
to the library.

Refs #295